Back-end/requirements.txt
# --- Django & REST ---
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-ratelimit==4.1.0
whitenoise==6.11.0
python-decouple==3.8
python-dotenv==1.1.1
asgiref==3.9.2
sqlparse==0.5.3

# --- LangChain ecosystem ---
langchain==1.0.3
langchain-core==1.0.2
langchain-community==0.4.1
langchain-groq==1.0.0
langchain-huggingface==1.0.0
langchain-text-splitters==1.0.0
langgraph==1.0.2

# --- Vector store & embeddings ---
chromadb==1.2.1
sentence-transformers==5.1.2
transformers==4.57.1
huggingface-hub==0.35.3
torch==2.9.0

# --- Utils & performance ---
pydantic==2.11.9
pydantic-settings==2.11.0
requests==2.32.5
numpy==2.2.6
scikit-learn==1.7.2
scipy==1.15.3
tqdm==4.67.1

# --- Auth & security ---
bcrypt==5.0.0
PyJWT==2.10.1

# --- Async server ---
uvicorn==0.38.0
httpx==0.28.1
anyio==4.11.0

# --- Logging & debug ---
rich==14.2.0
coloredlogs==15.0.1
humanfriendly==10.0

# --- Serialization & data handling ---
orjson==3.11.3
protobuf==6.33.0
PyYAML==6.0.3

# --- Optional GPU packages (comment out if CPU-only) ---
# nvidia-cublas-cu12==12.8.4.1
# nvidia-cuda-runtime-cu12==12.8.90
# nvidia-cudnn-cu12==9.10.2.21

# --- Testing ---
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-mock==3.14.0
factory-boy==3.3.1
faker==33.1.0
